Rename everything from picker to autocomplete
We decided to not change the name of the function itself for now as
that is a breaking change
