+++ docs/location-autocomplete-general.gif
