+++ scripts/prettify-autocomplete-graph.js
@@ -1,4 +1,4 @@
-const graph = require('../tmp/location-picker-graph.json')
+const graph = require('../tmp/location-autocomplete-graph.json')
 const stringify = require('json-stable-stringify')
 
 console.log(stringify(graph, { space: '  ' }))
